+++ temp_package/out/services/project-analyzer.js
@@ -1,0 +1,12 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.ProjectAnalyzer = void 0;
+// ProjectAnalyzer: 获取项目上下文信息
+class ProjectAnalyzer {
+    async getProjectContext() {
+        // TODO: 分析项目上下文
+        return { projectName: 'DemoProject', language: 'typescript' };
+    }
+}
+exports.ProjectAnalyzer = ProjectAnalyzer;
+//# sourceMappingURL=project-analyzer.js.map